Drawing shapes

+++ src/main/java/com/project1/MyPoint.java
@@ -13,7 +13,7 @@ public class MyPoint {
     //Constructor
     MyPoint() { setPoint(0,0); }
 
-    MyPoint(int x, int y)
+    MyPoint(double x, double y)
     {
         setPoint(x,y);
         setColor(pointColor);
@@ -33,7 +33,7 @@ public class MyPoint {
         this.y = p.getYCoordinate();
     }
 
-    public void setPoint(int x, int y) {
+    public void setPoint(double x, double y) {
         this.x = x;
         this.y = y;
     }
@@ -52,7 +52,7 @@ public class MyPoint {
     public double getYCoordinate(){ return y; }          //returns the y coordinate of the point
 
 
-    public void translate(int dx, int dy)                //Change the x and y coordinates by the specified distance
+    public void translate(double dx, double dy)                //Change the x and y coordinates by the specified distance
     {
         this.x += dx;
         this.y += dy;
